python/pyproject.toml
[project]
name = "kaleidoswap-sdk"
version = "0.2.0"
description = "Python SDK for KaleidoSwap - A decentralized exchange for Bitcoin and RGB assets"
authors = [
    {name = "Kaleidoswap Team", email = "dev@kaleidoswap.com"}
]
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["kaleidoswap", "bitcoin", "rgb", "lightning-network", "dex", "crypto", "trading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "mypy>=1.0.0",
    "ruff>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/kaleidoswap/kaleido-sdk"
Documentation = "https://github.com/kaleidoswap/kaleido-sdk#readme"
Repository = "https://github.com/kaleidoswap/kaleido-sdk.git"
Issues = "https://github.com/kaleidoswap/kaleido-sdk/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kaleidoswap_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = "-v --tb=short"
